third-party/BDWatchdog/MetricsFeeder/requirements.txt
python-daemon
requests